pyproject.toml
# beara_bones: Django site (itsbillw). Production: DietPi + MariaDB; dev: SQLite.
[project]
requires-python = ">= 3.12"
name = "beara_bones"
version = "0.1.0"
description = "Django based playground"
readme = "README.md"
dependencies = [
    "certbot-nginx>=2.11.0",
    "certbot>=2.11.0",
    "dbt-duckdb>=1.9.0",
    "django>=5.1.2",
    "duckdb>=1.0.0",
    "gunicorn>=23.0.0",
    "minio>=7.2.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "PyMySQL>=1.1.0",  # MariaDB driver (no native build on Pi); see beara_bones/__init__.py
    "python-dotenv>=1.0.1",
    "pyarrow>=18.0.0",
    "requests>=2.32.0",
    "soda-core-duckdb>=3.0.0",
    "soda-core>=3.0.0",
    "uvicorn-worker>=0.2.0",
    "uvicorn>=0.32.0",
    "whitenoise>=6.8.2",
]

[project.optional-dependencies]
dev = [
    "bandit>=1.7.0",
    "coverage[toml]>=7.0.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyMySQL>=1.1.0",
    "types-requests>=2.32.0",
]

[tool.setuptools.packages.find]
include = ["beara_bones*", "football*"]

[tool.coverage.run]
source = ["beara_bones"]
omit = ["*/migrations/*", "*/tests.py", "*/test_*.py", "manage.py"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]
fail_under = 0

[tool.bandit]
skips = ["B101", "B404", "B603", "B607"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
